Fuse fmpz_mpoly/mul*
